Get rid of duplicated compile tests
